+++ tests/testthat/test-node_get.R
@@ -1,18 +1,116 @@
 library(sfdep)
 library(dplyr)
+library(tidygraph)
 library(sfnetworks)
 
-net <- as_sfnetwork(roxel, directed = FALSE)
+# conflicted::conflict_prefer("activate", "sfnetworks")
+
+# create object with edge list and neighbor list
+net <- as_sfnetwork(roxel, directed = FALSE) |>
+  tidygraph::activate("edges") |>
+  mutate(len = edge_length()) |>
+  tidygraph::activate("nodes") |>
+  mutate(
+    nb = node_get_nbs(),
+    elist = node_get_edge_list(),
+    wt = node_get_edge_col(elist, "len")
+    )
+
+
+node_df <- as_tibble(net)
+
+edge_df <- net %E>%
+  as_tibble()
+
+# to test -----------------------------------------------------------------
+
+# attributes of nb class
+# attributes of wt columns such as style
+# check lengths on nbs and weights
+# check edge lists match weight list values
+
+
+# tests -------------------------------------------------------------------
+
+
+
+test_that("nbs is spdep compliant", {
+  nb <- as_tibble(net)[["nb"]]
+
+  # test that nb does not include self
+  expect_equal(attr(nb, "self.included"), FALSE)
+
+  # that nb is a nb class & list class
+  expect_s3_class(nb, c("nb", "list"))
+
+})
+
+
 
 test_that("identical adjacency lists", {
 
-  net_nb <- mutate(net, nb = node_get_nbs())
+  nb <- as_tibble(net)[["nb"]]
   ig_adj_list <- igraph::get.adjlist(net)
 
   expect_equal(
-    as_tibble(net_nb)$nb,
-    lapply(ig_adj_list, as.integer)
+    nb,
+    lapply(ig_adj_list, as.integer),
+    ignore_attr = TRUE
   )
 
+})
+
+
+test_that("edge list col is identical to igraph", {
+  n2 <- mutate(net, edge_list = node_get_edge_list())
+  el <- as_tibble(n2)[["edge_list"]]
+
+  elig <- lapply(igraph::get.adjedgelist(net), as.integer)
+
+  expect_equal(el, elig)
+})
+
+
+test_that("attr edge col has same lengths as nb", {
+  expect_equal(
+    lengths(node_df[["nb"]]),
+    lengths(node_df[["wt"]])
+    )
+})
+
+
+test_that("edge list contains correct neighbor nodes", {
+  nb <- node_df[["nb"]]
+  elist <- node_df[["elist"]]
+
+  correct_indexes <- logical(length(nb))
+
+  for (i in seq_along(elist)) {
+    tmp <- dplyr::slice(as_tibble(edge_df), elist[[i]])
+    res <- all(nb[[i]] %in% c(tmp[["from"]], tmp[["to"]]))
+    correct_indexes[i] <- res
+  }
+
+  expect_true(all(correct_indexes))
+
+})
+
+
+test_that("edge_get_col retrieves correct indexes", {
+  # check that edge list correctly matches neighbors
+  # the ith observation should all be from i to others
+  elist <- node_df[["elist"]]
+
+  # test that adjacency matrix matches edge list
+  # each element in the edge list should include `i` either in
+  # from or to as many times as there are neighbors
+  adj_vec <- integer(length(elist))
+  for (i in seq_along(elist)) {
+    tmp <- dplyr::slice(as_tibble(edge_df), elist[[i]])
+    res <- sum(c(tmp[["from"]], tmp[["to"]]) == i)
+    adj_vec[i] <- res
+  }
+
+  expect_equal(lengths(elist), adj_vec)
 })